Api tracking code cleanup
