+++ docker/docker-compose.yml
@@ -34,15 +34,6 @@ services:
       - DATABASE_NAME=${DATABASE_NAME}
       - DATABASE_USER=${DATABASE_USER}
       - DATABASE_PASSWORD=${DATABASE_PASSWORD}
-      #- WALLET_TYPE=${WALLET_TYPE}
-      #- POSTGRESQL_WALLET_HOST=${POSTGRESQL_WALLET_HOST}
-      #- POSTGRESQL_WALLET_PORT=${POSTGRESQL_WALLET_PORT}
-      #- POSTGRESQL_WALLET_USER=${POSTGRESQL_USER}
-      #- POSTGRESQL_WALLET_PASSWORD=${POSTGRESQL_PASSWORD}
-      #- POSTGRESQL_WALLET_ADMIN_PASSWORD=${POSTGRESQL_ADMIN_PASSWORD}
-      #- WALLET_ENCRYPTION_KEY=${WALLET_ENCRYPTION_KEY}
-      #- INDY_WALLET_SEED=${INDY_WALLET_SEED}
-      #- INDY_WALLET_DID=${INDY_WALLET_DID}
       - DEBUG=${DEBUG}
       - DEMO_SITE=${DEMO_SITE}
       - SQL_DEBUG=${SQL_DEBUG}
@@ -76,6 +67,10 @@ services:
       - RABBITMQ_PASSWORD=${RABBITMQ_PASSWORD}
       - AGENT_ADMIN_URL=http://vcr-agent:${AGENT_ADMIN_PORT}
       - AGENT_ADMIN_API_KEY=${AGENT_ADMIN_API_KEY}
+      - API_VERSION_ROUTING_MIDDLEWARE=${API_VERSION_ROUTING_MIDDLEWARE}
+      - SP_TRACKING_APP_ID=${SP_TRACKING_APP_ID}
+      - SP_TRACKING_EMITTER=${SP_TRACKING_EMITTER}
+      - SP_TRACKING_EMITTER_PROTOCOL=${SP_TRACKING_EMITTER_PROTOCOL}
       - RECORD_TIMINGS=${RECORD_TIMINGS}
       - TRACE_EVENTS=${TRACE_EVENTS}
       - TRACE_PROOF_EVENTS=${TRACE_PROOF_EVENTS}
